+++ rr/ns.js
@@ -1,0 +1,22 @@
+
+const RR = require('./index')
+
+class NS extends RR {
+  constructor (opts) {
+    super(opts)
+
+    if (!this.fullyQualified('NS', 'address', opts.address)) return
+    if (!this.validHostname('NS', 'address', opts.address)) return
+    this.set('address', opts.address)
+  }
+
+  toBind () {
+    return `${this.get('name')} ${this.get('ttl')} ${this.get('class')}  NS  ${this.get('address')}\n`
+  }
+
+  toTinydns () {
+    return `&${this.get('name')}::${this.get('address')}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
+  }
+}
+
+module.exports = NS