+++ CHANGES.md
@@ -2,10 +2,19 @@
 #### 1.N.N - YYYY-MM-DD
 
 
+#### 0.2.3 - 2021-10-21
+
+- refactored classes into separate files
+- TXT, SOA, MX, CNAME, PTR, SRV: add toBind and toTinydns
+- SOA: add setMinimum
+- lib/tinydns: added escapeOct & packDomainName
+- PTR, SRV: added tests
+
+
 #### 0.2.2 - 2021-10-20
 
 - add tests/*
-- add toBind and toTinydns for A and AAAA
+- A, AAAA, add toBind and toTinydns()
 - add .release